👌 IMPROVE: add dispacher to Form search for set Hotels
